added some changes

+++ src/components/utils/inventoryItemsApi.js
@@ -3,6 +3,12 @@ import { api, getHeader, getToken, getHeaderForFormData } from "./AppFunction";
 
 export async function createInventoryItems(inventoryId,productId, quantity, condition){
     try{
+        if(
+            !inventoryId || !productId || isNaN(quantity) || parseFloat(quantity) <= 0 ||
+            isNaN(condition) || parseInt(condition) <= 0
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {inventoryId, productId, quantity:parseFloat(quantity),condition:parseInt(condition)};
         const response = await api.post(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/create/new-inventory-items`,requestBody,{
             headers:getHeader()
@@ -21,6 +27,13 @@ export async function createInventoryItems(inventoryId,productId, quantity, cond
 
 export async function updateInventoryItems(id,inventoryId,productId, quantity, condition ) {
     try{
+        if(
+            !id ||
+            !inventoryId || !productId || isNaN(quantity) || parseFloat(quantity) <= 0 ||
+            isNaN(condition) || parseInt(condition) <= 0
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {inventoryId, productId, quantity:parseFloat(quantity),condition:parseInt(condition)};
         const response = await api.put(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/update/${id}`,requestBody,{
             headers:getHeader()
@@ -39,6 +52,9 @@ export async function updateInventoryItems(id,inventoryId,productId, quantity, c
 
 export async function deleteInventoryItems(id){
     try{
+        if(!id){
+            throw new Error("Dati id za inventoryItems nije pronadjen");
+        }
         const response = await api.delete(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/delete/${id}`,{
             headers:getHeader()
         });
@@ -51,6 +67,9 @@ export async function deleteInventoryItems(id){
 
 export async function findOneById(id){
     try{
+        if(!id){
+            throw new Error("Dati id za inventoryItems nije pronadjen");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/get-one/${id}`,{
             headers:getHeader()
         });
@@ -75,6 +94,9 @@ export async function findAll(){
 
 export async function getByQuantity(quantity){
     try{
+        if(isNaN(quantity) || parseFloat(quantity) <= 0){
+            throw new Error("Quantity must be at least 0");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/by-quantity`,{
             params:{
                 quantity:parseFloat(quantity)
@@ -90,6 +112,9 @@ export async function getByQuantity(quantity){
 
 export async function getByCondition(itemCondition){
     try{
+        if(isNaN(itemCondition) || parseInt(itemCondition) <= 0){
+            throw new Error("ItemCondition mora biti pozitivan broj");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/by-condition`,{
             params:{
                 itemCondition:parseInt(itemCondition)
@@ -105,6 +130,9 @@ export async function getByCondition(itemCondition){
 
 export async function getByInventoryId(inventoryId){
     try{
+        if(!inventoryId){
+            throw new Error("Dati ID za Inventory nije pronadjen");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/by-inventory/${inventoryId}`,{
             headers:getHeader()
         });
@@ -117,6 +145,9 @@ export async function getByInventoryId(inventoryId){
 
 export async function getByProductId(productId){
     try{
+        if(!productId){
+            throw new Error("Dati ID za Product nije pronadjen");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/by-product/${productId}`,{
             headers:getHeader()
         });
@@ -129,6 +160,9 @@ export async function getByProductId(productId){
 
 export async function getByProductName(productName){
     try{
+        if(!productName || typeof productName !== "string" || productName.trim() === ""){
+            throw new Error("Naziv proizvoda nije pronadjen");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/by-product-name`,{
             params:{
                 productName
@@ -144,6 +178,9 @@ export async function getByProductName(productName){
 
 export async function findItemsWithDifference(threshold){
     try{
+        if(isNaN(threshold) || parseFloat(threshold) <= 0){
+            throw new Error("Razlicite stavke/items nisu pornadjeni");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/inventoryItems/find-by-threshold`,{
             params:{
                 threshold:parseFloat(threshold)
