+++ src/components/utils/optionApi.js
@@ -1,7 +1,17 @@
 import { api, getHeader, getToken, getHeaderForFormData } from "./AppFunction";
 
+const isOptionCategoryValid = ["GENDER","ROLE","STATUS","LANGUAGE","THEME"]; 
+
 export async function createOption(label,value,category,active){
     try{
+        if(
+            !label || typeof label !=="string" || label.trim()==="" ||
+            !value || typeof value !=="string" || value.trim()==="" ||
+            !isOptionCategoryValid.includes(category.toUpperCase()) ||
+            typeof active !=="boolean"
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {label,value,category: (category || "").toUpperCase(),active};
         const response = await api.post(`${import.meta.env.VITE_API_BASE_URL}/option/create`,requestBody,{
             headers:getHeader()
@@ -20,6 +30,15 @@ export async function createOption(label,value,category,active){
 
 export async function updateOption(id,label,value,category,active){
     try{
+        if(
+            !id ||
+            !label || typeof label !=="string" || label.trim()==="" ||
+            !value || typeof value !=="string" || value.trim()==="" ||
+            !isOptionCategoryValid.includes(category.toUpperCase()) ||
+            typeof active !=="boolean"
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {label,value,category: (category || "").toUpperCase(),active};
         const response = await api.put(`${import.meta.env.VITE_API_BASE_URL}/option/update/${id}`,requestBody,{
             headers:getHeader()
@@ -38,6 +57,9 @@ export async function updateOption(id,label,value,category,active){
 
 export async function deleteOption(id){
     try{
+        if(!id){
+            throw new Error("Dati ID za option nije pronadjen");
+        }
         const response = await api.delete(`${import.meta.env.VITE_API_BASE_URL}/option/delete/${id}`,{
             headers:getHeader()
         });
@@ -45,6 +67,21 @@ export async function deleteOption(id){
     }
     catch(error){
         handleApiError(error, "Greska prilikom brisanja");
+    }
+}
+
+export async function getOne(id){
+    try{
+        if(!id){
+            throw new Error("Dati ID za option nije pronadjen");
+        }
+        const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/option/get-one/${id}`,{
+            headers:getHeader()
+        });
+        return response.data;
+    }
+    catch(error){
+        handleApiError(error,"Greska prilikom dobaljvanja jednog option-a");
     }
 }
 
@@ -62,6 +99,9 @@ export async function getAll(){
 
 export async function getByCategory(category){
     try{
+        if(!isOptionCategoryValid.includes(category.toUpperCase())){
+            throw new Error("Data kategorija nije pronadjena");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/option/category/${category}`,{
             headers:getHeader()
         });
