+++ src/components/utils/permissionApi.js
@@ -1,7 +1,14 @@
 import { api, getHeader, getToken, getHeaderForFormData } from "./AppFunction";
 
+const isPermisionTypeValid = ["VIEW_INVENTORY","CREATE_ORDER","CREATE_STORAGE_EMPLOYEE","CREATE_STORAGE_MANAGER"];
+
 export async function cratePermission(permissionType){
     try{
+        if(
+            !isPermisionTypeValid.includes(permissionType.toUpperCase())
+        ){
+            throw new Error("Tip mora biti validan i popunjen");
+        }
         const requestBody = {permissionType: (permissionType || "").toUpperCase()};
         const response = await api.post(`${import.meta.env.VITE_API_BASE_URL}/permission/create`,requestBody,{
             headers:getHeader()
@@ -27,6 +34,9 @@ export async function getAll(){
 
 export async function deletePermission(id){
     try{
+        if(!id){
+            throw new Error("Dati ID za permisson nije pronadjen");
+        }
         const response = await api.delete(`${import.meta.env.VITE_API_BASE_URL}/permission/delete/${id}`,{
             headers:getHeader()
         });
@@ -39,6 +49,12 @@ export async function deletePermission(id){
 
 export async function updatePermission(id, permissionType){
     try{
+        if(
+            !id ||
+            !isPermisionTypeValid.includes(permissionType.toUpperCase())
+        ){
+            throw new Error("Tip mora biti validan i popunjen");
+        }
         const requestBody = {permissionType: (permissionType || "").toUpperCase()};
         const response = await api.put(`${import.meta.env.VITE_API_BASE_URL}/permission/update/${id}`,requestBody,{
             headers:getHeader()
@@ -52,6 +68,9 @@ export async function updatePermission(id, permissionType){
 
 export async function getPermissionById(id){
     try{
+        if(!id){
+            throw new Error("Dati ID za permision nije pronadjen");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/permission/get/${id}`,{
             headers:getHeader()
         });
